fix(music-module): patch play-dl

+++ packages/music-module/tests/bot.js
@@ -7,10 +7,10 @@ import { CocoaIntent } from "cocoa-discord/template";
 
 import { Client } from "discord.js";
 
-// import { Music } from "@leomotors/music-bot"
+// import { Music } from "@cocoa-discord/music-module"
 import { Music } from "../dist/index.js";
 
-// * A simple discord bot to E2E (manually) test this cog module
+// * A simple discord bot to E2E (manually) test this module
 // * Also minimum code required to fire the bot
 // * Good Example!
 
@@ -18,21 +18,24 @@ const client = new Client(new CocoaIntent().useGuild().useGuildVoice());
 const style = new EmbedStyle({
   author: "invoker",
   color: 0xd7f6fc,
-  footer: { text: "@leomotors/music-bot test mode" },
+  footer: { text: "@cocoa-discord/music-module test mode" },
 });
 
-const center = new SlashCenter(client, process.env.GUILD_IDS?.split(","));
+if (!process.env.GUILD_ID) {
+  throw new Error("GUILD_ID is not set");
+}
+
+const center = new SlashCenter(client, [process.env.GUILD_ID]);
 center.addModules(new Music(client, style));
 center.useHelpCommand(style);
-center.on("error", (name, err, ctx) => {
+center.on("error", async (name, err, ctx) => {
+  await ctx.channel?.send(`Error ${err}`);
+  await client.destroy();
+
   // * FOR DEBUG PURPOSES ONLY, Don't do this on production
-  client.destroy();
   setTimeout(() => {
     throw err;
-  }, 100);
-
-  // * Note: What you should do on Production
-  // await ctx.channel?.send(`Error ${err}`);
+  }, 10);
 });
 
 client.on("ready", (cli) => {
